+++ WalletWasabi/WebClients/Coinstamp/CoinstampExchangeRate.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace WalletWasabi.WebClients.BlockchainInfo
+{
+	public partial class CoinstampExchangeRateProvider
+	{
+		public class CoinstampExchangeRate
+		{
+			[JsonProperty(PropertyName = "bid")]
+			public decimal Rate { get; set; }
+		}
+	}
+}
